currency service is done

+++ app/api/schemas.py
@@ -1,6 +1,9 @@
+from decimal import Decimal
+
 from pydantic import BaseModel
+
 
 
 class Currency(BaseModel):
     name: str
-    value: float
+    value: Decimal
